DRILL-2031: Parquet bit column reader fix
Removes attempted optimized bit reader, left a comment where it was removed as a pointer back to this change set. For now the higher level Parquet interface will be used to ensure reading is correct, a fix of the optimized reader can be pursued at a later date if is is necessary.

Fixed two bugs in the test framework necessary to verify the fix against the complex parquet reader (which uses a similar interface, but it ensures that none of the other columns are read incorrectly).

Changed code sightly based on recent automatic reallocation merge. Interface was changed on the setSafe method of the value vectors to not return a boolean. Moved cast out of the loop to make sure it only happens once.