SerializedExecutor sketch for consideration.
Based on the EventProcessor, but recast as an Executor that accepts Runnables.
The idea would be to use this on RPC reception points to immediately hand off
work. Each Fragment would have a SerializedExecutor associated with it, serving
as a queue to deliver events to when the RPC thread receives them.